+++ src/components/elements/Submenu/component.spec.js
@@ -43,16 +43,7 @@ describe('<Submenu />', () => {
         pointing: 'top left',
         trigger: children,
         upward: false,
-      });
-    });
-
-    it('should get `isSelectedDisabled` to set the `value` to `-999`', () => {
-      const wrapper = getSubmenuDropdown({
-        isSelectedDisabled: true,
-      });
-
-      expectComponentToHaveProps(wrapper, {
-        value: -999,
+        value: null,
       });
     });