+++ src/components/inputs/Dropdown/component.js
@@ -81,9 +81,8 @@ export class Component extends PureComponent {
           selection
           upward={willOpenAbove}
         />
-        {!hasImages && label && (
-          <label onClick={() => this.handleOpen(true)}>{label}</label>
-        )}
+        {!hasImages &&
+          label && <label onClick={() => this.handleOpen(true)}>{label}</label>}
       </div>
     );
   }