+++ src/components/property-page-widgets/PropertyPageHero/component.spec.js
@@ -4,15 +4,17 @@ import { expectComponentToHaveDisplayName } from '@lodgify/enzyme-jest-expect-he
 
 import { ComponentWithResponsive as PropertyPageHero } from './component';
 
-const url = '🚞';
+const imageUrl = '🚞';
 
 const props = {
   activeNavigationItemIndex: 1,
+  headerLogoSizes: 'a load of logo sizes',
   headerLogoSrc: 'src',
+  headerLogoSrcSet: 'a load of logo src sets',
   headerLogoText: 'text',
   headerNavigationItems: [{ text: 'Home', href: '/' }],
   headerPrimaryCTA: { onClick: Function.prototype, text: 'Book now' },
-  images: [{ url, label: 'Entrance' }, { url, label: 'Kitchen' }],
+  images: [{ imageUrl, label: 'Entrance' }, { imageUrl, label: 'Kitchen' }],
   searchBarGuestsOptions: [{ text: '1', value: '1' }],
   searchBarLocationOptions: [{ text: '1', value: '1' }],
   searchBarModalHeadingText: '😹',