+++ src/components/property-page-widgets/PropertyPageSearchBar/component.js
@@ -2,8 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import { Button } from 'elements/Button';
-import { ShowOnMobile } from 'layout/ShowOnMobile';
-import { ShowOnDesktop } from 'layout/ShowOnDesktop';
+import { ShowOn } from 'layout/ShowOn';
 import { CHECK_OUR_AVAILABILITY } from 'utils/default-strings';
 
 import { SearchBar } from '../../general-widgets/SearchBar/index';
@@ -32,7 +31,7 @@ export const Component = ({
 
   return (
     <div className="property-page-searchbar">
-      <ShowOnDesktop>
+      <ShowOn computer widescreen>
         <SearchBar
           getIsDayBlocked={getIsDayBlocked}
           guestsOptions={guestsOptions}
@@ -45,9 +44,10 @@ export const Component = ({
             ...summaryProps,
           })}
         />
-      </ShowOnDesktop>
-      <ShowOnMobile>
+      </ShowOn>
+      <ShowOn mobile tablet>
         <SearchBar
+          getIsDayBlocked={getIsDayBlocked}
           guestsOptions={guestsOptions}
           isDisplayedAsModal
           isFixed={isFixed}
@@ -57,12 +57,13 @@ export const Component = ({
           })}
           modalTrigger={modalTrigger}
           onChangeInput={onChangeInput}
+          onSubmit={onSubmit}
           summaryElement={getSummaryMarkup({
             areOnlyNightPriceAndRatingDisplayed: true,
             ...summaryProps,
           })}
         />
-      </ShowOnMobile>
+      </ShowOn>
     </div>
   );
 };