+++ src/components/property-page-widgets/RoomType/utils/getRoomFeaturesMarkup.js
@@ -4,7 +4,7 @@ import { List } from 'semantic-ui-react';
 import { GridColumn } from 'layout/GridColumn';
 import { Icon } from 'elements/Icon';
 import { Paragraph } from 'typography/Paragraph';
-
+import { buildKeyFromStrings } from 'utils/build-key-from-strings';
 /**
  * @param  {boolean}   showIcons
  * @param  {Object[]}  features
@@ -18,7 +18,7 @@ export const getRoomFeaturesMarkup = (showIcons, features) => (
     horizontal
   >
     {features.map(({ labelText, iconName }, index) => (
-      <List.Item key={index}>
+      <List.Item key={buildKeyFromStrings(labelText, index)}>
         {showIcons ? (
           <Paragraph>{labelText}</Paragraph>
         ) : (