+++ src/index.js
@@ -60,8 +60,7 @@ export { Grid } from './components/layout/Grid';
 export { GridColumn } from './components/layout/GridColumn';
 export { GridRow } from './components/layout/GridRow';
 export { HorizontalGutters } from './components/layout/HorizontalGutters';
-export { ShowOnDesktop } from './components/layout/ShowOnDesktop';
-export { ShowOnMobile } from './components/layout/ShowOnMobile';
+export { ShowOn } from './components/layout/ShowOn';
 export { VerticalGutters } from './components/layout/VerticalGutters';
 
 // Media
@@ -80,8 +79,8 @@ export { HostProfile } from './components/property-page-widgets/HostProfile';
 export { KeyFacts } from './components/property-page-widgets/KeyFacts';
 export { Location } from './components/property-page-widgets/Location';
 export {
-  PaymentInformation,
-} from './components/property-page-widgets/PaymentInformation';
+  PolicyAndNotes,
+} from './components/property-page-widgets/PolicyAndNotes';
 export {
   PropertyPageHero,
 } from './components/property-page-widgets/PropertyPageHero';