+++ src/utils/build-key-from-strings/buildKeyFromStrings.js
@@ -3,4 +3,4 @@
  * @return {string}
  */
 export const buildKeyFromStrings = (...strings) =>
-  strings.map(string => string.toString()).join('');
+  strings.map((string = '') => string.toString()).join('');