+++ src/utils/build-key-from-strings/buildKeyFromStrings.spec.js
@@ -4,6 +4,7 @@ describe('buildKeyFromStrings', () => {
   it('should compose a key from an array of strings and/or integers', () => {
     const testCases = [
       { strings: ['yo', 1], expected: 'yo1' },
+      { strings: ['yo', undefined, 1], expected: 'yo1' },
       { strings: ['yo', 'yo', 2], expected: 'yoyo2' },
       { strings: ['yo', 'yo', 'yo', 3], expected: 'yoyoyo3' },
     ];